Steps
bower install was never mentioned, and Gulp doesn't appear to do that.

But also, perhaps it should?  What's the overall mgmt strategy here... should box install be comprehensive?  Or npm install? Or should gulp be able to update these dependencies?  Or is the workflow only to run bower when you need to?